IssueController 내 속성 지정 메서드 API Controller로 이관 및 테스트코드 수정
