+++ src/test/java/codesquad/web/ApiIssueControllerTest.java
@@ -12,6 +12,8 @@ import support.test.AcceptanceTest;
 import static codesquad.domain.IssueTest.*;
 import static codesquad.domain.UserTest.BRAD;
 import static codesquad.domain.UserTest.JUNGHYUN;
+import static codesquad.domain.label.LabelTest.LABEL;
+import static codesquad.domain.milestone.MilestoneTest.MILESTONE;
 import static org.slf4j.LoggerFactory.getLogger;
 
 public class ApiIssueControllerTest extends AcceptanceTest {
@@ -74,5 +76,23 @@ public class ApiIssueControllerTest extends AcceptanceTest {
         String location = createResource("/api/issues", BRAD, NEW_ISSUE_BODY);
         ResponseEntity<Void> responseEntity = basicAuthTemplate(JUNGHYUN).exchange(location, HttpMethod.DELETE, null, Void.class);
         softly.assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.FORBIDDEN);
+    }
+
+    @Test
+    public void setMilestone() {
+        ResponseEntity<String> response = template.getForEntity(String.format("/api/issues/%d/milestones/%d", ISSUE.getId(), MILESTONE.getId()), String.class);
+        softly.assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
+    }
+
+    @Test
+    public void setAssignee() {
+        ResponseEntity<String> response = template.getForEntity(String.format("/api/issues/%d/assignees/%d", ISSUE.getId(), BRAD.getId()), String.class);
+        softly.assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
+    }
+
+    @Test
+    public void setLabel() {
+        ResponseEntity<String> response = template.getForEntity(String.format("/api/issues/%d/labels/%d", ISSUE.getId(), LABEL.getId()), String.class);
+        softly.assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
     }
 }
