+++ src/test/java/codesquad/web/IssueAcceptanceTest.java
@@ -3,7 +3,6 @@ package codesquad.web;
 import codesquad.domain.issue.Issue;
 import codesquad.domain.issue.IssueRepository;
 import codesquad.domain.milestone.Milestone;
-import codesquad.service.MilestoneService;
 import org.junit.Before;
 import org.junit.Test;
 import org.slf4j.Logger;
@@ -18,8 +17,6 @@ import support.test.HtmlFormDataBuilder;
 import static codesquad.domain.IssueTest.*;
 import static codesquad.domain.UserTest.BRAD;
 import static codesquad.domain.UserTest.JUNGHYUN;
-import static codesquad.domain.label.LabelTest.LABEL;
-import static codesquad.domain.milestone.MilestoneTest.MILESTONE;
 import static codesquad.domain.milestone.MilestoneTest.MILESTONES;
 import static org.slf4j.LoggerFactory.getLogger;
 
@@ -183,26 +180,5 @@ public class IssueAcceptanceTest extends BasicAuthAcceptanceTest {
 
         ResponseEntity<String> response = template().postForEntity(createdIssue.generateUrl(), deleteRequest, String.class);
         softly.assertThat(response.getStatusCode()).isEqualTo(HttpStatus.UNAUTHORIZED);
-    }
-
-    @Test
-    public void setMilestone() {
-        ResponseEntity<String> response = template.getForEntity(String.format("/issues/%d/milestones/%d", ISSUE.getId(), MILESTONE.getId()), String.class);
-        softly.assertThat(response.getStatusCode()).isEqualTo(HttpStatus.FOUND);
-        softly.assertThat(response.getHeaders().getLocation().getPath()).isEqualTo(String.format("/issues/%d", ISSUE.getId()));
-    }
-
-    @Test
-    public void setAssignee() {
-        ResponseEntity<String> response = template.getForEntity(String.format("/issues/%d/assignees/%d", ISSUE.getId(), BRAD.getId()), String.class);
-        softly.assertThat(response.getStatusCode()).isEqualTo(HttpStatus.FOUND);
-        softly.assertThat(response.getHeaders().getLocation().getPath()).isEqualTo(String.format("/issues/%d", ISSUE.getId()));
-    }
-
-    @Test
-    public void setLabel() {
-        ResponseEntity<String> response = template.getForEntity(String.format("/issues/%d/labels/%d", ISSUE.getId(), LABEL.getId()), String.class);
-        softly.assertThat(response.getStatusCode()).isEqualTo(HttpStatus.FOUND);
-        softly.assertThat(response.getHeaders().getLocation().getPath()).isEqualTo(String.format("/issues/%d", ISSUE.getId()));
     }
 }
